requirements.txt
numpy==1.24.3
pandas==2.0.2
tensorflow==2.12.0
scikit-learn==1.2.2
matplotlib==3.7.1
opencv-python==4.7.0.72
Pillow==9.5.0
